fix :  [ user server ]  스케줄링 코드 수정
삭제할 계정이 존재하는 경우에만 삭제를 진행해야 하는데
스케줄링이 실행 될 때 마다 삭제 프로세스를 실행하고 있었다.

고침

issue key : SB2-6, SB2-7, #48

+++ simpleBoard02/server-b-user/src/main/java/kim/zhyun/serveruser/service/impl/UserSchedulerImpl.java
@@ -32,16 +32,17 @@ public class UserSchedulerImpl implements UserScheduler {
         Set<Long> userIds = deleteUserIdList(rdbList);
         Set<String> redisList = deleteListForRedis(rdbList);
         
-        log.info("📆 Scheduler start - delete count = RDB: {}, userIds: {}, Redis: {} ----┐", rdbList.size(), userIds.size(), redisList.size());
-        
-        var response = articleClient.withdrawalArticleDelete(userIds);
-        
-        if (response.getBody().getStatus()) {
-            userRepository.deleteAllInBatch(rdbList);
-            redisTemplate.delete(redisList);
+        if (rdbList.size() > 0) {
+            log.info("📆 Scheduler start - delete count = RDB: {}, userIds: {}, Redis: {} ----┐", rdbList.size(), userIds.size(), redisList.size());
+            var response = articleClient.withdrawalArticleDelete(userIds);
+            
+            if (response.getBody().getStatus()) {
+                userRepository.deleteAllInBatch(rdbList);
+                redisTemplate.delete(redisList);
+            }
+            
+            log.info("📆 Scheduler end--------------------------------------------------------┘");
         }
-        
-        log.info("📆 Scheduler end--------------------------------------------------------┘");
     }
     
     /**
